Don't panic, but reject promise if invalid JSON is returned by API
